refact(Balances): reorganize files
