+++ src/services/balances/api.js
@@ -1,4 +1,4 @@
-import { prepareURL, getApi } from './_helpers';
+import { prepareURL, getApi } from 'api/_helpers.js';
 
 const balancesURL = (orgId) => prepareURL(`/api/organizations/${orgId}/total_balances`);
